osd: shuffle UnstableHashInfoRegistry to ECCommon
