+++ src/osd/ECBackend.h
@@ -324,29 +324,7 @@ public:
 
   const ECUtil::stripe_info_t sinfo;
 
-  class UnstableHashInfoRegistry {
-    CephContext *cct;
-    ceph::ErasureCodeInterfaceRef ec_impl;
-    /// If modified, ensure that the ref is held until the update is applied
-    SharedPtrRegistry<hobject_t, ECUtil::HashInfo> registry;
-
-  public:
-    UnstableHashInfoRegistry(
-      CephContext *cct,
-      ceph::ErasureCodeInterfaceRef ec_impl)
-      : cct(cct),
-	ec_impl(std::move(ec_impl)) {}
-
-    ECUtil::HashInfoRef maybe_put_hash_info(
-      const hobject_t &hoid,
-      ECUtil::HashInfo &&hinfo);
-
-    ECUtil::HashInfoRef get_hash_info(
-      const hobject_t &hoid,
-      bool create,
-      const std::map<std::string, ceph::buffer::list, std::less<>>& attr,
-      uint64_t size);
-  } unstable_hashinfo_registry;
+  ECCommon::UnstableHashInfoRegistry unstable_hashinfo_registry;
 
   int object_stat(const hobject_t &hoid, struct stat* st);
